% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.R
\name{assemble}
\alias{assemble}
\title{Assemble your charm bracelet}
\usage{
assemble(path = here("_charm.R"))
}
\arguments{
\item{path}{The path to a _charm.R file.}
}
\description{
Assemble your charm bracelet
}
