% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charm.R
\name{charm}
\alias{charm}
\title{Create a charm}
\usage{
charm(goal, ingredients = NULL, instructions)
}
\arguments{
\item{goal}{A file path.}

\item{ingredients}{A vector of file paths (optional).}

\item{instructions}{A vector of either paths to R files or R commands, which
are strings that begin with a lowercase r and are followed by a space and
an R command. Like inline R code in R Markdown.}
}
\description{
Create a charm
}
