% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_slate.R
\name{clean_slate}
\alias{clean_slate}
\title{Start your charm bracelet from scratch}
\usage{
clean_slate(path = here("_charm.R"))
}
\arguments{
\item{path}{The path to a _charm.R file.}
}
\description{
Start your charm bracelet from scratch
}
