% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mise_en_place.R
\name{mise_en_place}
\alias{mise_en_place}
\title{Set up your charm bracelet}
\usage{
mise_en_place(path = here())
}
\arguments{
\item{path}{A path to a directory where you want to start your charm bracelet.}
}
\description{
Set up your charm bracelet
}
